feature/15 (#17)
* 画像保存ジョブ化

* index付与

* bigIntegerに変更

* 物理削除バッチ

* cronファイルに改行追加

* Dockerfile修正

* supervisord設定

* 不要ログ削除

+++ database/migrations/2025_04_27_015503_create_m_menus_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('m_menus', function (Blueprint $table) {
             $table->id()->comment('献立ID');
-            $table->integer('users_id')->comment('ユーザーID');
-            $table->integer('recipes_id')->comment('レシピID');
+            $table->bigInteger('users_id')->comment('ユーザーID');
+            $table->bigInteger('recipes_id')->comment('レシピID');
             $table->date('date')->comment('日');
             $table->integer('time_zone_type')->comment('時間帯ID');
             $table->text('memo')->nullable()->comment('メモ');
